refactor: 클래스명의 버전 삭제 및 파일 이동
