+++ tomcat/src/main/java/org/apache/catalina/startup/Controllers.java
@@ -11,7 +11,7 @@ public class Controllers {
     public Controllers(HttpController... controller) {
         value = Arrays.stream(controller).toList();
     }
-    
+
     boolean contains(String path) {
         return value.stream().anyMatch(r -> r.getPath().equals(path));
     }
