+++ tomcat/src/main/java/org/apache/coyote/Processor.java
@@ -24,9 +24,8 @@ import java.net.Socket;
 public interface Processor {
 
     /**
-     * Process a connection. This is called whenever an event occurs (e.g. more
-     * data arrives) that allows processing to continue for a connection that is
-     * not currently being processed.
+     * Process a connection. This is called whenever an event occurs (e.g. more data arrives) that allows processing to
+     * continue for a connection that is not currently being processed.
      */
     void process(Socket socket);
 }
