--- tomcat/src/main/java/org/apache/coyote/http11/request/Http11Request.java
@@ -1,10 +1,0 @@
-package org.apache.coyote.http11.request;
-
-import java.util.List;
-
-public final class Http11Request extends HttpRequest {
-    public Http11Request(List<String> clientData) {
-        super(clientData);
-    }
-
-}
