+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpReader.java
@@ -7,14 +7,14 @@ import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Http11Reader extends BufferedReader {
-    public Http11Reader(InputStream inputStream) {
+public class HttpReader extends BufferedReader {
+    public HttpReader(InputStream inputStream) {
         super(new InputStreamReader(inputStream));
     }
 
     public List<String> readLines() throws IOException {
         List<String> result = new ArrayList<>(readLinesWhileBuffered());
-        if(ready()) {
+        if (ready()) {
             result.add("");
             result.add(readWhileBuffered());
         }
