+++ tomcat/src/main/java/org/apache/coyote/http11/request/requestline/HttpVersion.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.requestline;
 
 import java.util.Arrays;
