+++ tomcat/src/test/java/org/apache/coyote/http11/request/HttpHeadersTest.java
@@ -1,7 +1,5 @@
 package org.apache.coyote.http11.request;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
