+++ tomcat/src/test/java/org/apache/coyote/http11/request/HttpLocationTest.java
@@ -1,5 +1,6 @@
 package org.apache.coyote.http11.request;
 
+import org.apache.coyote.http11.request.requestline.HttpLocation;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
