+++ tomcat/src/test/java/org/apache/coyote/http11/request/HttpMethodTest.java
@@ -1,6 +1,7 @@
 package org.apache.coyote.http11.request;
 
 
+import org.apache.coyote.http11.request.requestline.HttpMethod;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
